+++ src/main/java/com/lab2/backend/repository/DecanatoRepository.java
@@ -3,7 +3,7 @@ import com.lab2.backend.model.Decanato;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface DecanatoRepository extends JpaRepository<Decanato, Integer>{
+public interface DecanatoRepository extends JpaRepository<Decanato, Long>{
     List<Decanato> findByEstatus(Character estatus);
 
 }
